+++ src/Domain/Budget/Command/SubtractAmountFromBudget.php
@@ -7,7 +7,7 @@ class SubtractAmountFromBudget
     private $budgetId;
     private $amount;
 
-    public function __construct(string $budgetId, int $amount)
+    public function __construct(string $budgetId, float $amount)
     {
         $this->budgetId = $budgetId;
         $this->amount = $amount;
@@ -18,7 +18,7 @@ class SubtractAmountFromBudget
         return $this->budgetId;
     }
 
-    public function getAmount(): int
+    public function getAmount(): float
     {
         return $this->amount;
     }
